tiles and linear maps for cats
